requirements.txt
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
greenlet==1.1.0
inventorize3==0.0.1
joblib==1.0.1
numpy==1.21.1
pandas==1.3.1
prettytable==2.1.0
Pygments==2.9.0
pyodbc==4.0.31
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
rich==10.7.0
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
SQLAlchemy==1.4.22
threadpoolctl==2.2.0
wcwidth==0.2.5
